+++ src/ha_mcp/tools/config_entry_flow.py
@@ -14,20 +14,27 @@ helpers: ``ha_set_integration`` creates entries for arbitrary domains through
 ``create_config_entry`` and edits them through ``update_config_entry_options``,
 and ``ha_config_set_helper(helper_type="config_subentry")`` drives subentry
 flows through ``set_config_subentry``.
+
+The step machinery those entry points drive lives in three sibling modules,
+imported in one direction only (menu <- form <- walker <- here):
+
+- ``config_entry_flow_menu``: menu selection keys and menu-step handling
+- ``config_entry_flow_form``: form-step schema consumption and reuse tracking
+- ``config_entry_flow_walker``: step submission, HA error translation, flow
+  introspection, and the two flow walkers
 """
 
 import asyncio
-import copy
-import json
 import logging
-from collections.abc import Iterator
-from dataclasses import dataclass
-from dataclasses import field as dc_field
-from enum import StrEnum
-from typing import Any, Literal, NoReturn
+from typing import Any, Literal
 
-from ..client.rest_client import HomeAssistantAPIError
 from ..errors import ErrorCode, create_error_response
+from .config_entry_flow_form import _extract_schema_field_names
+from .config_entry_flow_walker import (
+    _FlowType,
+    _handle_config_subentry_flow_steps,
+    _handle_flow_steps,
+)
 from .helpers import raise_tool_error
 
 logger = logging.getLogger(__name__)
@@ -73,40 +80,6 @@ FLOW_HELPER_TYPES: frozenset[str] = frozenset(
     }
 )
 
-# Keys used to specify a menu selection — stripped before submitting form data.
-# The tuple fixes the multi-key precedence (frozenset iteration order is
-# hash-randomized per process, so consuming in set order would make a config
-# carrying two selection keys nondeterministic); the frozenset serves the
-# membership checks. Mirrored by ``_MENU_CHOICE_CONFIG_KEYS`` in
-# tools_config_helpers.py.
-_MENU_SELECTION_KEY_ORDER = ("group_type", "next_step_id", "menu_option")
-_MENU_SELECTION_KEYS = frozenset(_MENU_SELECTION_KEY_ORDER)
-
-# Flow step types an MCP client cannot drive: external steps need a browser
-# (OAuth / cloud authorization), progress steps wait on HA-side async work.
-# Surfaced as structured errors instead of attempted (issue #1814).
-_UNDRIVABLE_STEP_TYPES = frozenset(
-    {"external", "external_done", "progress", "progress_done"}
-)
-_RECONFIGURE_SUCCESS_REASONS = frozenset(
-    {
-        "reauth_successful",
-        "reconfigure_successful",
-    }
-)
-_MISSING_DEFAULT = object()
-# "Submit nothing for this field" — see _redeclared_field_submission.
-_NO_SUBMISSION: tuple[Any, bool] = (_MISSING_DEFAULT, False)
-
-
-class _FlowType(StrEnum):
-    """HA config flow result type strings."""
-
-    FORM = "form"
-    MENU = "menu"
-    ABORT = "abort"
-    CREATE_ENTRY = "create_entry"
-
 
 # ---------------------------------------------------------------------------
 # Module-level flow machinery
@@ -115,1505 +88,6 @@ class _FlowType(StrEnum):
 # tools_config_helpers.py. They take a client instance as an explicit
 # parameter so the same logic can be used from any caller.
 # ---------------------------------------------------------------------------
-
-
-def _handle_menu_step(
-    flow_id: str,
-    current_step: dict[str, Any],
-    remaining_config: dict[str, Any],
-    consumed_selections: list[str] | None = None,
-    consumed_selection_keys: list[str] | None = None,
-) -> str:
-    """Extract menu selection from config, raising on missing selection.
-
-    Returns the menu choice string. Mutates remaining_config: a scalar
-    selection is popped outright; a list of successive selections yields one
-    element per menu encounter, the un-consumed tail replacing the key until
-    it runs dry. Lists let flows that revisit a menu (issue #2116:
-    battery_sim loops menu → branch form → menu until 'all_done') be walked
-    within the step budget the walkers derive from the selection count. The
-    caller's list object is never mutated.
-
-    ``consumed_selections`` (when provided by the walker) accumulates every
-    selection consumed during the walk, so a menu revisited after the
-    selections ran dry can raise an error that names what was already
-    consumed instead of claiming no selection was supplied;
-    ``consumed_selection_keys`` records which selection key served each one,
-    so that error's example uses the caller's own key.
-    """
-    menu_choice = None
-    for key in _MENU_SELECTION_KEY_ORDER:
-        if key not in remaining_config:
-            continue
-        value = remaining_config[key]
-        if isinstance(value, list):
-            if not value:
-                # Empty list = no selection supplied under this key; another
-                # selection key may still carry one.
-                remaining_config.pop(key)
-                continue
-            menu_choice = value[0]
-            if not menu_choice:
-                # A falsy element mid-list is a malformed call, not an
-                # exhausted list — without this check it would raise the
-                # "ran dry" error while later selections sit unconsumed,
-                # pointing the caller at a problem they don't have.
-                raise_tool_error(
-                    create_error_response(
-                        ErrorCode.VALIDATION_INVALID_PARAMETER,
-                        f"Menu selection list under '{key}' contains an "
-                        f"empty element ({menu_choice!r}) — every element "
-                        "must be a menu option name.",
-                        suggestions=[
-                            f"Available options: "
-                            f"{current_step.get('menu_options', [])}",
-                        ],
-                        context={
-                            "flow_id": flow_id,
-                            "step_id": current_step.get("step_id"),
-                            "selection_key": key,
-                        },
-                    )
-                )
-            rest = list(value[1:])
-            if rest:
-                remaining_config[key] = rest
-            else:
-                remaining_config.pop(key)
-        else:
-            menu_choice = remaining_config.pop(key)
-        break
-
-    if not menu_choice:
-        _raise_missing_menu_selection(
-            flow_id, current_step, consumed_selections, consumed_selection_keys
-        )
-
-    choice = str(menu_choice)
-    if consumed_selections is not None:
-        consumed_selections.append(choice)
-    if consumed_selection_keys is not None:
-        consumed_selection_keys.append(key)
-    return choice
-
-
-def _raise_missing_menu_selection(
-    flow_id: str,
-    current_step: dict[str, Any],
-    consumed_selections: list[str] | None,
-    consumed_selection_keys: list[str] | None,
-) -> NoReturn:
-    """Raise the no-selection error for a menu step.
-
-    Distinguishes a revisited menu whose selections ran dry (name what was
-    consumed, show the list syntax under the caller's own key) from a first
-    menu given no selection at all (the original guidance).
-    """
-    menu_options = current_step.get("menu_options", [])
-    context = {
-        "flow_id": flow_id,
-        "step_id": current_step.get("step_id"),
-        "menu_options": menu_options,
-    }
-    if consumed_selections:
-        # "<next-selection>" is a placeholder on purpose: cyclic flows
-        # routinely need the SAME option again (battery_sim's create flow
-        # picks 'no_tariff_info' once per meter), so no concrete option
-        # can be suggested without guessing the caller's path.
-        example = json.dumps([*consumed_selections, "<next-selection>"])
-        example_key = (
-            consumed_selection_keys[-1] if consumed_selection_keys else "next_step_id"
-        )
-        raise_tool_error(
-            create_error_response(
-                ErrorCode.CONFIG_MISSING_REQUIRED_FIELDS,
-                "Flow presented another menu after the supplied "
-                f"selection(s) ({', '.join(map(repr, consumed_selections))}) "
-                "were consumed. Pass your menu selection key "
-                "('next_step_id', 'group_type', or 'menu_option') as a "
-                "list of successive selections to drive flows that "
-                "revisit menus.",
-                suggestions=[
-                    f"Available options: {menu_options}",
-                    f'Example: {{"{example_key}": {example}}}',
-                ],
-                context={
-                    **context,
-                    "consumed_menu_selections": list(consumed_selections),
-                },
-            )
-        )
-    raise_tool_error(
-        create_error_response(
-            ErrorCode.CONFIG_MISSING_REQUIRED_FIELDS,
-            "Menu step requires a selection. "
-            "Add 'group_type' or 'next_step_id' to your config.",
-            suggestions=[
-                f"Available options: {menu_options}",
-                'Example: {"group_type": "light", "name": "My Group", ...}',
-            ],
-            context=context,
-        )
-    )
-
-
-def _flow_step_budget(config: dict[str, Any]) -> int:
-    """Step allowance for one flow walk.
-
-    The historical cap of 10 guarded against runaway flows, but a cyclic
-    flow driven by an N-item selection list legitimately costs about two
-    steps per selection (the menu plus the branch step it opens) —
-    battery_sim's create flow spends 8 of the 10 on a minimal two-meter
-    setup, and one fixed-tariff branch more would starve it. Grant every
-    caller-supplied selection its two steps on top of the base; the budget
-    stays bounded by the caller's own input, so the runaway protection is
-    intact.
-    """
-    selections = 0
-    for key in _MENU_SELECTION_KEY_ORDER:
-        value = config.get(key)
-        if isinstance(value, list):
-            selections += len(value)
-        elif value is not None:
-            selections += 1
-    return 10 + 2 * selections
-
-
-def iter_schema_fields(data_schema: Any) -> Iterator[dict[str, Any]]:
-    """Yield leaf field definitions from a flow schema, descending into sections."""
-    if not isinstance(data_schema, list):
-        return
-    for field in data_schema:
-        if not isinstance(field, dict):
-            continue
-        nested_schema = field.get("schema")
-        if isinstance(nested_schema, list):
-            yield from iter_schema_fields(nested_schema)
-            continue
-        yield field
-
-
-def _section_path(path_prefix: str, name: Any) -> str:
-    """Return the dotted config path for a named or anonymous section."""
-    if not isinstance(name, str):
-        return path_prefix
-    return f"{path_prefix}.{name}" if path_prefix else name
-
-
-@dataclass
-class _ReuseState:
-    """Caller-supplied form values a flow's steps have already consumed.
-
-    A config key applies to every form step that declares it, not only the
-    first step to pop it out of ``remaining_config`` — see
-    :func:`_redeclared_field_submission` for when a recorded value is
-    resubmitted. The record is split by how the caller wrote the value:
-
-    - ``scoped`` maps a dotted declaration path to a value that came out of an
-      explicitly supplied section dict. The caller named that section, so the
-      value belongs to that path and nowhere else.
-    - ``flat`` maps a leaf name to a value that came out of the flat caller
-      dict. The caller named no section, so the value is position-agnostic and
-      fills that leaf wherever a later step declares it.
-
-    ``filled`` holds the dotted paths the current step already filled from the
-    caller's own keys, so nothing is injected over a value the caller wrote for
-    this very step; it is cleared per step. ``fired`` bounds resubmission to
-    one write per (step, path) for the whole flow, ``notes`` collects the
-    warning each of those writes emits, and ``step_id`` names the step being
-    consumed.
-    """
-
-    scoped: dict[str, Any] = dc_field(default_factory=dict)
-    flat: dict[str, Any] = dc_field(default_factory=dict)
-    filled: set[str] = dc_field(default_factory=set)
-    fired: set[str] = dc_field(default_factory=set)
-    notes: list[str] = dc_field(default_factory=list)
-    step_id: str | None = None
-
-    def begin_step(self, step_id: Any) -> None:
-        """Start recording for a new form step."""
-        self.step_id = step_id if isinstance(step_id, str) else None
-        self.filled.clear()
-
-    def record(
-        self, path_prefix: str, name: str, value: Any, *, scoped_only: bool
-    ) -> None:
-        """Snapshot a value the caller supplied at this declaration site.
-
-        A flat value that lands on a path already recorded from an explicit
-        section dict also replaces that scoped entry: the flat key is the one
-        actually submitted for the path (flat overrides explicit), so the
-        record must carry the effective value or a later redeclaration of the
-        path would resurrect the overridden one.
-        """
-        dotted = _section_path(path_prefix, name)
-        self.filled.add(dotted)
-        if scoped_only:
-            self.scoped[dotted] = copy.deepcopy(value)
-        else:
-            self.flat[name] = copy.deepcopy(value)
-            if dotted in self.scoped:
-                self.scoped[dotted] = copy.deepcopy(value)
-
-    def recorded_value(self, path_prefix: str, name: str) -> Any:
-        """Return the value recorded for this declaration site, else _MISSING_DEFAULT."""
-        dotted = _section_path(path_prefix, name)
-        if dotted in self.scoped:
-            return copy.deepcopy(self.scoped[dotted])
-        if name in self.flat:
-            return copy.deepcopy(self.flat[name])
-        return _MISSING_DEFAULT
-
-    def claim_write(self, dotted: str) -> bool:
-        """Spend the single resubmission allowed for this (step, path), if unspent.
-
-        A flow that re-presents the same step gets one reused write and then
-        HA's own loud "required key not provided" naming the field, rather than
-        an unbounded run of silent rewrites.
-        """
-        key = f"{self.step_id}:{dotted}"
-        if key in self.fired:
-            return False
-        self.fired.add(key)
-        self.notes.append(
-            f"Resubmitted '{dotted}' at step '{self.step_id}': supplied once "
-            "but requested by more than one step encounter in this flow "
-            "(a later step redeclaring the field, or the same step revisited "
-            "via a menu loop — per-visit values cannot be expressed)"
-        )
-        return True
-
-
-def _record_ignored_section_keys(
-    ignored_config_keys: set[str] | None,
-    remaining_config: dict[str, Any],
-    section_path: str,
-) -> None:
-    """Record undeclared keys remaining inside an explicit section dict."""
-    if ignored_config_keys is None:
-        return
-    ignored_config_keys.update(
-        f"{section_path}.{key}" if section_path else key
-        for key in remaining_config
-        if key not in _MENU_SELECTION_KEYS
-    )
-
-
-def _field_default_value(field: dict[str, Any]) -> Any:
-    """Return a serialized schema field's default/suggested value, if present."""
-    description = field.get("description")
-    if isinstance(description, dict) and description.get("suggested_value") is not None:
-        return copy.deepcopy(description["suggested_value"])
-    if field.get("suggested_value") is not None:
-        return copy.deepcopy(field["suggested_value"])
-    if "default" in field:
-        return copy.deepcopy(field["default"])
-    return _MISSING_DEFAULT
-
-
-def _schema_default_values(data_schema: list[Any]) -> dict[str, Any]:
-    """Build default form data from serialized schema suggestions/defaults."""
-    defaults: dict[str, Any] = {}
-    for field in data_schema:
-        if not isinstance(field, dict):
-            continue
-        name = field.get("name")
-        nested_schema = field.get("schema")
-        if not isinstance(name, str):
-            continue
-        if isinstance(nested_schema, list):
-            nested_defaults = _schema_default_values(nested_schema)
-            if nested_defaults:
-                defaults[name] = nested_defaults
-            continue
-        default = _field_default_value(field)
-        if default is not _MISSING_DEFAULT:
-            defaults[name] = default
-    return defaults
-
-
-def _required_section_defaults(field: dict[str, Any]) -> dict[str, Any]:
-    """Return default data for a required section, otherwise an empty dict."""
-    nested_schema = field.get("schema")
-    if not field.get("required") or not isinstance(nested_schema, list):
-        return {}
-    return _schema_default_values(nested_schema)
-
-
-def _ignored_keys_warnings(
-    ignored_config_keys: set[str], remaining_config: dict[str, Any]
-) -> list[str]:
-    """Build warnings for caller-supplied config keys no flow step consumed."""
-    warnings: list[str] = []
-    ignored = ignored_config_keys | {
-        key for key in remaining_config if key not in _MENU_SELECTION_KEYS
-    }
-    if ignored:
-        warnings.append(
-            "Ignored config keys not declared by the Home Assistant flow "
-            f"schema: {', '.join(sorted(ignored))}"
-        )
-    leftover_menu_keys = _MENU_SELECTION_KEYS & remaining_config.keys()
-    if leftover_menu_keys:
-        # Name the values, not just the keys: an un-consumed selection means
-        # that branch was never configured, which a success response would
-        # otherwise hide (the scalar era carried no information beyond the
-        # key name; a list does).
-        details = ", ".join(
-            f"{key}={remaining_config[key]!r}" for key in sorted(leftover_menu_keys)
-        )
-        warnings.append(
-            f"Ignored menu selection key(s) with no matching menu step: {details}"
-        )
-    return warnings
-
-
-def _success_warnings(
-    ignored_config_keys: set[str],
-    remaining_config: dict[str, Any],
-    reuse_state: _ReuseState,
-) -> list[str]:
-    """Build a success response's ``warnings`` list (empty when there is nothing to say).
-
-    Merges the keys no step declared with the resubmissions the walk performed,
-    keeping ``warnings`` a flat ``list[str]`` per the response contract in
-    ``tests/src/unit/test_helper_response_shape.py``.
-    """
-    return _ignored_keys_warnings(ignored_config_keys, remaining_config) + list(
-        reuse_state.notes
-    )
-
-
-def _consume_section_schema(
-    field: dict[str, Any],
-    explicit_section: dict[str, Any] | None,
-    remaining_config: dict[str, Any],
-    ignored_config_keys: set[str] | None,
-    consumed_config_keys: set[str] | None,
-    path_prefix: str,
-    reuse_state: _ReuseState | None = None,
-    *,
-    allow_reuse: bool = True,
-    explicit_source: bool = False,
-) -> dict[str, Any]:
-    """Consume config values for a nested flow section.
-
-    Values inside an explicitly supplied section dict are consumed first, then
-    flat caller keys, which is what lets a flat child override the same key
-    written inside the section dict.
-    """
-    nested_schema = field.get("schema")
-    if not isinstance(nested_schema, list):
-        return {}
-
-    name = field.get("name")
-    section_path = _section_path(path_prefix, name)
-    nested_data = _required_section_defaults(field)
-
-    if explicit_section is not None:
-        explicit_remaining = dict(explicit_section)
-        nested_data.update(
-            _consume_form_schema(
-                nested_schema,
-                explicit_remaining,
-                ignored_config_keys,
-                consumed_config_keys,
-                section_path,
-                reuse_state,
-                allow_reuse=allow_reuse,
-                explicit_source=True,
-            )
-        )
-        _record_ignored_section_keys(
-            ignored_config_keys,
-            explicit_remaining,
-            section_path,
-        )
-
-    nested_data.update(
-        _consume_form_schema(
-            nested_schema,
-            remaining_config,
-            ignored_config_keys,
-            consumed_config_keys,
-            section_path,
-            reuse_state,
-            allow_reuse=allow_reuse,
-            explicit_source=explicit_source,
-        )
-    )
-    return nested_data
-
-
-def _mark_consumed(
-    consumed_config_keys: set[str] | None,
-    path_prefix: str,
-    name: str,
-) -> None:
-    """Record that a caller-supplied value was used, fresh or resubmitted.
-
-    Values the step's own schema supplies — section defaults, suggestions,
-    constants — are never marked: they are HA's data, and marking them would
-    let a config of nothing but misspelled keys look partially applied to
-    :func:`_finish_flow_entry`.
-    """
-    if consumed_config_keys is not None:
-        consumed_config_keys.add(_section_path(path_prefix, name))
-
-
-def _auto_confirm_form_payload(current_step: dict[str, Any]) -> dict[str, Any] | None:
-    """Return payload for HA preview/confirmation-only forms we can safely advance."""
-    if "preview" not in current_step:
-        return None
-    data_schema = current_step.get("data_schema")
-    if not isinstance(data_schema, list) or len(data_schema) != 1:
-        return None
-    field = data_schema[0]
-    if not isinstance(field, dict) or not field.get("required"):
-        return None
-    name = field.get("name")
-    if not isinstance(name, str) or name in _MENU_SELECTION_KEYS:
-        return None
-    default = _field_default_value(field)
-    if default is not False:
-        return None
-    selector = field.get("selector")
-    if isinstance(selector, dict) and selector and "boolean" not in selector:
-        return None
-    return {name: True}
-
-
-def _step_owned_submission_value(field: dict[str, Any]) -> Any:
-    """Return the value a step's own schema supplies for ``field``, else _MISSING_DEFAULT.
-
-    Deliberately distinct from :func:`_field_default_value`, which answers
-    "what would the UI show in this box" and is what seeds a form. This answers
-    "what does the step itself say to submit", which is a different question:
-
-    - ``voluptuous_serialize`` emits ``"default"`` only for an actual
-      voluptuous default. HA's edit-style pre-fill
-      (``add_suggested_values_to_schema``) copies the marker and overwrites
-      only its description, so a marker that already carried a default
-      serializes with both keys; the suggestion is the stored current value
-      and outranks the static default.
-    - A constant field serializes as ``{"type": "constant", "value": X}`` and
-      ``X`` is the only value it accepts.
-
-    The bare top-level ``suggested_value`` shape is not something
-    ``voluptuous_serialize`` produces; it is read defensively alongside the
-    nested one.
-    """
-    description = field.get("description")
-    if isinstance(description, dict) and description.get("suggested_value") is not None:
-        return copy.deepcopy(description["suggested_value"])
-    if field.get("suggested_value") is not None:
-        return copy.deepcopy(field["suggested_value"])
-    if field.get("type") == "constant" and field.get("value") is not None:
-        return copy.deepcopy(field["value"])
-    return _MISSING_DEFAULT
-
-
-def _redeclared_field_submission(
-    field: dict[str, Any],
-    name: str,
-    path_prefix: str,
-    reuse_state: _ReuseState | None,
-    allow_reuse: bool,
-) -> tuple[Any, bool]:
-    """Decide what to submit for a declared field the caller named no key for here.
-
-    Returns ``(value, from_caller)``, or ``_NO_SUBMISSION`` to omit the key
-    entirely. A site the caller's own key already filled earlier in this same
-    step is left alone — a section dict and a flat key can name the same leaf,
-    and the caller's value for this step outranks anything injected. Otherwise,
-    in order:
-
-    1. The field is not required: omit it. Nothing is ever injected into an
-       optional field, or into a section that is neither required nor named by
-       the caller (``allow_reuse``) — materializing either would invent data.
-    2. The step's own schema supplies a value — a suggestion or a constant's
-       only legal value, per :func:`_step_owned_submission_value`: submit that.
-       It is schema data rather than a caller key, so it is neither marked
-       consumed nor warned about. A suggestion outranks a coexisting
-       ``"default"``: both keys can serialize together, and the suggestion is
-       the stored current value while the default is the static schema one —
-       omitting would let voluptuous substitute the static value over it.
-    3. The field carries a ``"default"`` key (and no value of its own): omit it
-       and let voluptuous fill the default in. Key presence is the test, so
-       ``default: None`` is a default too.
-    4. Otherwise the field is required, has no default and has no value of its
-       own, which makes omitting it a guaranteed "required key not provided":
-       resubmit the value the caller supplied for an earlier step, warn, and
-       spend the one write allowed per (step, path).
-
-    Mutates ``reuse_state`` on the fourth branch. Menu selection keys never
-    reach here. Motivating regression (issue #2057): an options flow —
-    LocalTuya's — that declares the same field on an early step and again on a
-    later one.
-    """
-    if reuse_state is None or not allow_reuse:
-        return _NO_SUBMISSION
-    dotted = _section_path(path_prefix, name)
-    if dotted in reuse_state.filled:
-        return _NO_SUBMISSION
-    if not field.get("required"):
-        return _NO_SUBMISSION
-    step_owned = _step_owned_submission_value(field)
-    if step_owned is not _MISSING_DEFAULT:
-        return step_owned, False
-    if "default" in field:
-        return _NO_SUBMISSION
-    recorded = reuse_state.recorded_value(path_prefix, name)
-    if recorded is _MISSING_DEFAULT:
-        return _NO_SUBMISSION
-    if not reuse_state.claim_write(dotted):
-        return _NO_SUBMISSION
-    return recorded, True
-
-
-def _consume_leaf_field(
-    field: dict[str, Any],
-    name: str,
-    form_data: dict[str, Any],
-    remaining_config: dict[str, Any],
-    consumed_config_keys: set[str] | None,
-    reuse_state: _ReuseState | None,
-    path_prefix: str,
-    *,
-    allow_reuse: bool = True,
-    explicit_source: bool = False,
-) -> None:
-    """Fill ``name`` in ``form_data`` from the caller's config or from the step itself.
-
-    A key the caller supplied here is popped, submitted, and recorded in
-    ``reuse_state`` — scoped to this dotted path when it came out of an
-    explicitly supplied section dict, keyed by leaf name when it came out of
-    the flat caller dict. With no key to pop,
-    :func:`_redeclared_field_submission` chooses between omitting the field,
-    submitting the step's own value, and resubmitting the recorded one.
-    """
-    if name in remaining_config:
-        value = remaining_config.pop(name)
-        form_data[name] = value
-        _mark_consumed(consumed_config_keys, path_prefix, name)
-        if reuse_state is not None:
-            reuse_state.record(path_prefix, name, value, scoped_only=explicit_source)
-        return
-
-    value, from_caller = _redeclared_field_submission(
-        field, name, path_prefix, reuse_state, allow_reuse
-    )
-    if value is _MISSING_DEFAULT:
-        return
-    form_data[name] = value
-    if from_caller:
-        _mark_consumed(consumed_config_keys, path_prefix, name)
-
-
-def _consume_declared_section(
-    field: dict[str, Any],
-    form_data: dict[str, Any],
-    remaining_config: dict[str, Any],
-    ignored_config_keys: set[str] | None,
-    consumed_config_keys: set[str] | None,
-    path_prefix: str,
-    reuse_state: _ReuseState | None,
-    *,
-    allow_reuse: bool,
-    explicit_source: bool,
-) -> None:
-    """Merge one section field's data into ``form_data``.
-
-    A caller who supplies the section as a non-dict value gets it submitted
-    verbatim — HA, not this walker, decides what that means. Reuse is allowed
-    inside the section only when HA marks it required or the caller named it,
-    so an untouched optional section is never brought into existence.
-    """
-    name = field.get("name")
-    section_name = name if isinstance(name, str) else None
-    explicit_section: dict[str, Any] | None = None
-    if section_name is not None and section_name in remaining_config:
-        explicit_value = remaining_config.pop(section_name)
-        if not isinstance(explicit_value, dict):
-            form_data[section_name] = explicit_value
-            _mark_consumed(consumed_config_keys, path_prefix, section_name)
-            return
-        explicit_section = explicit_value
-
-    nested_data = _consume_section_schema(
-        field,
-        explicit_section,
-        remaining_config,
-        ignored_config_keys,
-        consumed_config_keys,
-        path_prefix,
-        reuse_state,
-        allow_reuse=allow_reuse
-        and (bool(field.get("required")) or explicit_section is not None),
-        explicit_source=explicit_source,
-    )
-    if not nested_data:
-        return
-    if section_name is not None:
-        form_data[section_name] = nested_data
-    else:
-        form_data.update(nested_data)
-
-
-def _consume_form_schema(
-    data_schema: list[Any],
-    remaining_config: dict[str, Any],
-    ignored_config_keys: set[str] | None = None,
-    consumed_config_keys: set[str] | None = None,
-    path_prefix: str = "",
-    reuse_state: _ReuseState | None = None,
-    *,
-    allow_reuse: bool = True,
-    explicit_source: bool = False,
-) -> dict[str, Any]:
-    """Consume matching config values and shape nested flow sections.
-
-    Mutates ``remaining_config`` by removing every consumed key. Flat child
-    values override the same value inside an explicitly supplied section dict.
-    Unknown keys inside explicit section dicts are added to
-    ``ignored_config_keys`` with their dotted section path. A declared field the
-    caller named no key for is filled per
-    :func:`_redeclared_field_submission`.
-    """
-    form_data: dict[str, Any] = {}
-
-    for field in data_schema:
-        if not isinstance(field, dict):
-            continue
-
-        name = field.get("name")
-        if isinstance(field.get("schema"), list):
-            _consume_declared_section(
-                field,
-                form_data,
-                remaining_config,
-                ignored_config_keys,
-                consumed_config_keys,
-                path_prefix,
-                reuse_state,
-                allow_reuse=allow_reuse,
-                explicit_source=explicit_source,
-            )
-            continue
-
-        if isinstance(name, str) and name not in _MENU_SELECTION_KEYS:
-            _consume_leaf_field(
-                field,
-                name,
-                form_data,
-                remaining_config,
-                consumed_config_keys,
-                reuse_state,
-                path_prefix,
-                allow_reuse=allow_reuse,
-                explicit_source=explicit_source,
-            )
-
-    return form_data
-
-
-def _extract_schema_field_names(data_schema: Any) -> set[str] | None:
-    """Extract the set of field names declared by a step's data_schema.
-
-    HA returns data_schema as a list of {name, selector, required, ...} dicts.
-    Nested leaf names are included; section-container names are omitted.
-    Returns ``None`` when the schema is absent or not a list (signalling
-    the caller to fall back to legacy submit-all behaviour). Returns a
-    (possibly empty) set when the schema is present and parseable.
-    """
-    if not isinstance(data_schema, list):
-        return None
-    names: set[str] = set()
-    for field in iter_schema_fields(data_schema):
-        name = field.get("name")
-        if isinstance(name, str):
-            names.add(name)
-    return names
-
-
-def _consume_all_remaining_keys(
-    remaining_config: dict[str, Any],
-    consumed_config_keys: set[str] | None,
-    reuse_state: _ReuseState | None,
-) -> dict[str, Any]:
-    """Submit every non-menu key, for a step whose schema HA did not send.
-
-    Without field names there is nothing to filter on, so the whole config is
-    dumped into this one submit and cleared out of ``remaining_config``. Each
-    consumed key is still recorded by leaf name, so a later step that *does*
-    arrive with a schema declaring one of them can be filled from the record
-    rather than submitted without it.
-
-    In a cyclic-menu flow the config legitimately carries values destined
-    for several branches, so a schemaless step mid-loop sweeping them all up
-    is far likelier to misfire ("extra keys not allowed" attributed to the
-    wrong step) — a warning is surfaced when this fires while menu
-    selections are still queued, so the sweep is at least visible.
-    """
-    form_data: dict[str, Any] = {}
-    for key in list(remaining_config.keys()):
-        if key in _MENU_SELECTION_KEYS:
-            continue
-        value = remaining_config.pop(key)
-        form_data[key] = value
-        _mark_consumed(consumed_config_keys, "", key)
-        if reuse_state is not None:
-            reuse_state.record("", key, value, scoped_only=False)
-    queued = {
-        key: remaining_config[key]
-        for key in _MENU_SELECTION_KEY_ORDER
-        if key in remaining_config
-    }
-    if form_data and queued and reuse_state is not None:
-        reuse_state.notes.append(
-            "A step without a schema consumed every remaining config key "
-            f"({', '.join(sorted(form_data))}) while menu selection(s) "
-            f"{queued!r} were still queued — values intended for later "
-            "branches may have been submitted to this step instead"
-        )
-    return form_data
-
-
-def _handle_form_step(
-    flow_id: str,
-    current_step: dict[str, Any],
-    remaining_config: dict[str, Any],
-    ignored_config_keys: set[str] | None = None,
-    consumed_config_keys: set[str] | None = None,
-    reuse_state: _ReuseState | None = None,
-) -> dict[str, Any]:
-    """Validate a form step and return form data to submit.
-
-    When the step's ``data_schema`` is provided, pops ONLY the keys declared
-    in that schema from ``remaining_config`` (mutating it) so any unconsumed
-    keys remain available for subsequent steps. Menu selection keys are never
-    submitted. Fields declared inside a section are grouped under the section
-    key; callers may provide them flat or inside an explicit section dict.
-
-    Caller-supplied values are recorded in ``reuse_state``. A field this step
-    declares but the caller named no key for here is filled per
-    :func:`_redeclared_field_submission`: omitted when the schema carries a
-    ``"default"`` or the field is optional, submitted from the step's own
-    suggestion or constant, and otherwise — required, no default, no value of
-    its own — resubmitted once from an earlier step's caller value with a
-    warning. Nothing is injected into an optional field, or into a section
-    neither marked required nor named by the caller.
-
-    When ``data_schema`` is absent (HA didn't tell us field names), falls
-    back to legacy behaviour: submit all non-menu keys and clear them. This
-    keeps single-step flows working when HA omits the schema.
-
-    Raises ToolError on validation errors.
-    """
-    if current_step.get("errors"):
-        raise_tool_error(
-            create_error_response(
-                ErrorCode.VALIDATION_INVALID_PARAMETER,
-                "Form validation failed",
-                suggestions=["Fix the field errors and retry with corrected values"],
-                context={
-                    "flow_id": flow_id,
-                    "step_id": current_step.get("step_id"),
-                    "errors": current_step["errors"],
-                    "data_schema": current_step.get("data_schema"),
-                },
-            )
-        )
-
-    if reuse_state is not None:
-        reuse_state.begin_step(current_step.get("step_id"))
-
-    data_schema = current_step.get("data_schema")
-    if not isinstance(data_schema, list):
-        return _consume_all_remaining_keys(
-            remaining_config, consumed_config_keys, reuse_state
-        )
-
-    return _consume_form_schema(
-        data_schema,
-        remaining_config,
-        ignored_config_keys,
-        consumed_config_keys,
-        "",
-        reuse_state,
-    )
-
-
-def _parse_flow_api_error(
-    api_error: HomeAssistantAPIError,
-) -> dict[str, Any]:
-    """Extract structured field-level info from an HA flow 4xx response.
-
-    Home Assistant returns voluptuous validation failures during flow
-    submission as either:
-
-    - ``{"message": "User input malformed: extra keys not allowed @ data['name']"}``
-      (raised before form validation, e.g. unknown field in payload)
-    - ``{"errors": {"base": "..."}, "description_placeholders": {...}}``
-      (per-field errors after voluptuous validation succeeds)
-    - Free-form text (when the body isn't JSON).
-
-    Returns a dict with at least:
-      - ``message``: the most informative human-readable string we found.
-      - ``field_errors``: dict of field-name -> error code/message, when
-        the body contained an ``errors`` map. Empty dict otherwise.
-      - ``raw``: the response_data dict (or ``None``) for diagnostics.
-    """
-    body = api_error.response_data or {}
-    field_errors: dict[str, Any] = {}
-    message_parts: list[str] = []
-
-    if isinstance(body, dict):
-        errors_field = body.get("errors")
-        if isinstance(errors_field, dict):
-            field_errors = {
-                key: val for key, val in errors_field.items() if isinstance(key, str)
-            }
-
-        # HA's stock 400 carries a `message` key with the voluptuous detail.
-        msg = body.get("message")
-        if isinstance(msg, str) and msg.strip():
-            message_parts.append(msg.strip())
-
-        # description_placeholders sometimes carry the human-readable error.
-        placeholders = body.get("description_placeholders")
-        if isinstance(placeholders, dict):
-            for key, val in placeholders.items():
-                if isinstance(val, str) and val.strip():
-                    message_parts.append(f"{key}: {val.strip()}")
-
-    if not message_parts:
-        # Fall back to the wrapper exception message ("API error: 400 - ...").
-        message_parts.append(str(api_error))
-
-    return {
-        "message": " | ".join(dict.fromkeys(message_parts)),  # de-dupe, preserve order
-        "field_errors": field_errors,
-        "raw": body if isinstance(body, dict) else None,
-    }
-
-
-async def _process_menu_flow_result(
-    flow_result: dict[str, Any],
-    client: Any,
-    intro_flow_id: str | None,
-    menu_choice: str | None,
-) -> dict[str, Any]:
-    """Return schema or menu_options dict for a MENU-type flow result."""
-    info: dict[str, Any] = {}
-    if menu_choice:
-        if not intro_flow_id:
-            return info
-        try:
-            step = await asyncio.wait_for(
-                client.submit_config_flow_step(
-                    intro_flow_id, {"next_step_id": menu_choice}
-                ),
-                timeout=10.0,
-            )
-        except (HomeAssistantAPIError, TimeoutError):
-            return info
-        if step.get("type") == _FlowType.FORM:
-            schema = step.get("data_schema")
-            if isinstance(schema, list):
-                info["schema"] = schema
-        return info
-
-    options = flow_result.get("menu_options")
-    if isinstance(options, list):
-        filtered = [opt for opt in options if isinstance(opt, str)]
-        if filtered:
-            info["menu_options"] = filtered
-    return info
-
-
-async def fetch_helper_flow_info(
-    client: Any,
-    helper_type: str | None,
-    menu_choice: str | None = None,
-) -> dict[str, Any]:
-    """Best-effort introspection of a helper's config-entry flow.
-
-    Starts a fresh introspection flow (always aborted) and returns a dict
-    with optional keys ``"schema"`` and ``"menu_options"`` so a single HA
-    round-trip serves both the schema-attach path (used by
-    ``_raise_flow_api_error`` and the pre-flow validation gates in
-    ``_handle_flow_helper``) and the menu-sub-types path (used when a
-    menu-rooted helper has no branch chosen yet — issue #1186).
-
-    Behaviour:
-
-    - FORM at top: ``{"schema": [...]}``
-    - MENU at top with ``menu_choice``: submits and returns the branch
-      form schema as ``{"schema": [...]}`` (no ``menu_options`` since
-      the caller already picked a branch)
-    - MENU at top without ``menu_choice``: ``{"menu_options": [...]}``
-    - any failure or unparseable shape: ``{}`` (callers branch on
-      ``"schema" in info`` / ``"menu_options" in info``)
-    """
-    info: dict[str, Any] = {}
-    if not helper_type or client is None:
-        return info
-    intro_flow_id: str | None = None
-    try:
-        flow_result = await client.start_config_flow(helper_type)
-        intro_flow_id = flow_result.get("flow_id")
-        flow_type = flow_result.get("type")
-
-        if flow_type == _FlowType.FORM:
-            schema = flow_result.get("data_schema")
-            if isinstance(schema, list):
-                info["schema"] = schema
-            return info
-
-        if flow_type == _FlowType.MENU:
-            return await _process_menu_flow_result(
-                flow_result, client, intro_flow_id, menu_choice
-            )
-
-        return info
-    except Exception:
-        return info
-    finally:
-        if intro_flow_id:
-            try:
-                await asyncio.wait_for(
-                    client.abort_config_flow(intro_flow_id), timeout=5.0
-                )
-            except Exception as abort_err:
-                logger.debug(
-                    f"Failed to abort introspection flow {intro_flow_id}: {abort_err}"
-                )
-
-
-def _build_flow_error_context(
-    flow_id: str,
-    status_code: int,
-    helper_type: str | None,
-    menu_choice: str | None,
-    current_step: dict[str, Any] | None,
-    submitted: dict[str, Any] | None,
-    parsed_raw: dict[str, Any] | None,
-) -> dict[str, Any]:
-    context: dict[str, Any] = {"flow_id": flow_id, "status_code": status_code}
-    if helper_type:
-        context["helper_type"] = helper_type
-    if menu_choice:
-        context["menu_choice"] = menu_choice
-    if current_step is not None:
-        context["step_id"] = current_step.get("step_id")
-    if submitted is not None:
-        context["submitted_keys"] = sorted(submitted.keys())
-    if parsed_raw is not None:
-        context["response_body"] = parsed_raw
-    return context
-
-
-async def _raise_flow_api_error(
-    api_error: HomeAssistantAPIError,
-    *,
-    client: Any,
-    flow_id: str,
-    helper_type: str | None,
-    menu_choice: str | None,
-    current_step: dict[str, Any] | None,
-    submitted: dict[str, Any] | None,
-) -> None:
-    """Translate an HA 4xx during a flow submit into a structured ToolError.
-
-    For 400/422 responses, parses ``response_data`` for field-level info
-    via ``_parse_flow_api_error``. When the body is unstructured (no
-    ``errors`` map), attaches the helper's ``data_schema`` (if it can be
-    fetched) so the caller has actionable information.
-
-    Always raises ``ToolError`` — never returns.
-    """
-    parsed = _parse_flow_api_error(api_error)
-    field_errors = parsed["field_errors"]
-    status_code = api_error.status_code or 0
-
-    context = _build_flow_error_context(
-        flow_id,
-        status_code,
-        helper_type,
-        menu_choice,
-        current_step,
-        submitted,
-        parsed["raw"],
-    )
-
-    suggestions: list[str] = []
-    message: str
-
-    current_schema = None
-    if current_step is not None:
-        step_schema = current_step.get("data_schema")
-        if isinstance(step_schema, list):
-            current_schema = step_schema
-
-    # Single introspection round-trip — used by both branches below.
-    info = await fetch_helper_flow_info(client, helper_type, menu_choice)
-    schema = info.get("schema") or current_schema
-
-    if field_errors:
-        # Structured field errors — tell the caller which fields failed.
-        context["field_errors"] = field_errors
-        readable = ", ".join(f"{k}: {v}" for k, v in field_errors.items())
-        message = f"Helper validation failed — {readable}"
-        suggestions.append(
-            "Fix the field(s) listed in 'field_errors' and retry the call."
-        )
-        # Issue #1149: also attach the data_schema so the LLM sees the field
-        # shape (selector, required, ...) alongside the per-field error
-        # codes — symmetric with the unstructured-error branch below.
-        # `field_errors` tells "what failed", `data_schema` tells "what's
-        # accepted"; together they're enough for self-correction.
-        if schema is not None:
-            context["data_schema"] = schema
-    else:
-        # Unstructured — attach the data_schema so the LLM has something to use.
-        message = (
-            f"Home Assistant rejected the {helper_type or 'flow'} request "
-            f"({status_code}): {parsed['message']}"
-        )
-        if schema is not None:
-            context["data_schema"] = schema
-            suggestions.append(
-                "Inspect 'data_schema' in this error to see the fields HA expects, "
-                "then retry with a corrected config."
-            )
-
-    raise_tool_error(
-        create_error_response(
-            ErrorCode.SERVICE_CALL_FAILED,
-            message,
-            suggestions=suggestions,
-            context=context,
-        )
-    )
-
-
-async def _submit_step(
-    submit_fn: Any,
-    flow_id: str,
-    payload: dict[str, Any],
-    *,
-    client: Any,
-    helper_type: str | None,
-    last_menu_choice: str | None,
-    current_step: dict[str, Any],
-) -> dict[str, Any]:
-    try:
-        return await asyncio.wait_for(submit_fn(flow_id, payload), timeout=20.0)
-    except HomeAssistantAPIError as api_err:
-        if api_err.status_code in (400, 422):
-            await _raise_flow_api_error(
-                api_err,
-                client=client,
-                flow_id=flow_id,
-                helper_type=helper_type,
-                menu_choice=last_menu_choice,
-                current_step=current_step,
-                submitted=payload,
-            )
-        raise
-
-
-def _finish_flow_entry(
-    flow_id: str,
-    current_step: dict[str, Any],
-    *,
-    supplied_keys: list[str],
-    saw_form_step: bool,
-    any_form_key_consumed: bool,
-    ignored_config_keys: set[str],
-    remaining_config: dict[str, Any],
-    reuse_state: _ReuseState,
-) -> dict[str, Any]:
-    """Build the CREATE_ENTRY success response, or raise on total key miss.
-
-    When the flow presented at least one form step, the caller supplied
-    config keys, and NONE were consumed by any form (typos / wrong field
-    names), the flow was walked on empty forms and HA saved form defaults —
-    not the caller's values. A success + warning there reads as "done" to an
-    LLM caller, so fail loudly with the schema route instead. Partial
-    consumption — and flows that complete without any form step (instant
-    creates) — keep the established success + warnings contract.
-    """
-    if supplied_keys and saw_form_step and not any_form_key_consumed:
-        raise_tool_error(
-            create_error_response(
-                ErrorCode.VALIDATION_INVALID_PARAMETER,
-                "Flow completed without consuming any of the supplied "
-                "config keys — every form step was submitted empty, so "
-                "the flow saved its defaults, not your values",
-                suggestions=[
-                    "Check the field names against the flow's data_schema — "
-                    "ha_get_integration(entry_id=..., include_schema=True) "
-                    "shows the accepted fields — then retry with corrected "
-                    "keys.",
-                ],
-                context={
-                    "flow_id": flow_id,
-                    "supplied_keys": supplied_keys,
-                    "details": current_step,
-                },
-            )
-        )
-    response: dict[str, Any] = {"success": True, "entry": current_step}
-    warnings = _success_warnings(ignored_config_keys, remaining_config, reuse_state)
-    if warnings:
-        response["warnings"] = warnings
-    return response
-
-
-def _raise_flow_abort(flow_id: str, current_step: dict[str, Any]) -> NoReturn:
-    """Raise the structured error for an ABORT flow step."""
-    reason = current_step.get("reason")
-    abort_suggestions: list[str] = []
-    if reason in ("already_configured", "single_instance_allowed"):
-        # Common benign aborts on the add-integration path (#1814): give the
-        # caller a route to the existing entry instead of a bare failure.
-        abort_suggestions.append(
-            "The integration is already set up — use "
-            "ha_get_integration() to find the existing entry."
-        )
-    raise_tool_error(
-        create_error_response(
-            ErrorCode.SERVICE_CALL_FAILED,
-            f"Flow aborted: {reason}",
-            suggestions=abort_suggestions or None,
-            context={"flow_id": flow_id, "details": current_step},
-        )
-    )
-
-
-async def _handle_flow_steps(
-    client: Any,
-    flow_id: str,
-    initial_step: dict[str, Any],
-    config: dict[str, Any],
-    submit_fn: Any = None,
-    helper_type: str | None = None,
-) -> dict[str, Any]:
-    """Walk a multi-step config flow handling menu and form steps (max 10 steps).
-
-    HA flows can present steps in sequence:
-    - ``menu``: caller supplies selection via ``group_type``/``next_step_id`` key
-    - ``form``: caller supplies field values; aborts immediately on validation errors
-    - ``create_entry``: flow complete
-    - ``abort``: flow terminated by HA
-
-    Args:
-        client: HomeAssistantClient instance
-        flow_id: Flow ID from start_config_flow or start_options_flow
-        initial_step: The first step returned by the flow start call
-        config: Full caller-provided config dict. Menu selection keys are
-            consumed by menu steps — a key may carry a single selection or a
-            list of successive selections, consumed one per menu encounter
-            (flows can revisit a menu after each branch; issue #2116).
-            Remaining keys are submitted on the first
-            form step whose schema declares them or, when HA omits the schema,
-            on the first form step outright. A later step that redeclares an
-            already-consumed field gets that value resubmitted once, with a
-            warning, where the step marks it required and supplies neither a
-            default nor a value of its own (see
-            :func:`_redeclared_field_submission`).
-        submit_fn: Async function to submit a step. Defaults to
-            client.submit_config_flow_step (create). Pass
-            client.submit_options_flow_step for options (update) flows.
-        helper_type: Optional helper type (e.g. ``"statistics"``). When
-            provided, surfaces the helper's data_schema in error context
-            for unstructured HA 4xx responses so the caller can react.
-
-    Returns:
-        ``{"success": True, "entry": result}`` on success, plus ``warnings``
-        when SOME caller-supplied config keys were not declared by any flow
-        step, or when a key had to be resubmitted to a later step that
-        redeclared it. When the flow presented at least one form step and NONE
-        of the supplied keys were consumed, raises ``VALIDATION_INVALID_PARAMETER``
-        instead of reporting a misleading success — the flow completed on
-        empty forms (defaults), applying nothing the caller asked for (see
-        :func:`_finish_flow_entry`). Raises ToolError on any failure.
-    """
-    if submit_fn is None:
-        submit_fn = client.submit_config_flow_step
-    remaining_config = dict(config)
-    current_step = initial_step
-    last_menu_choice: str | None = None
-    consumed_menu_selections: list[str] = []
-    consumed_menu_selection_keys: list[str] = []
-    ignored_config_keys: set[str] = set()
-    reuse_state = _ReuseState()
-    supplied_keys = sorted(k for k in config if k not in _MENU_SELECTION_KEYS)
-    saw_form_step = False
-    any_form_key_consumed = False
-    max_steps = _flow_step_budget(config)
-
-    for step_num in range(max_steps):
-        result_type = current_step.get("type")
-
-        if result_type == _FlowType.CREATE_ENTRY:
-            return _finish_flow_entry(
-                flow_id,
-                current_step,
-                supplied_keys=supplied_keys,
-                saw_form_step=saw_form_step,
-                any_form_key_consumed=any_form_key_consumed,
-                ignored_config_keys=ignored_config_keys,
-                remaining_config=remaining_config,
-                reuse_state=reuse_state,
-            )
-
-        if result_type == _FlowType.ABORT:
-            _raise_flow_abort(flow_id, current_step)
-
-        if result_type == _FlowType.MENU:
-            menu_choice = _handle_menu_step(
-                flow_id,
-                current_step,
-                remaining_config,
-                consumed_menu_selections,
-                consumed_menu_selection_keys,
-            )
-            last_menu_choice = menu_choice
-            logger.debug(
-                f"Flow step {step_num}: menu '{menu_choice}' "
-                f"(step_id={current_step.get('step_id')})"
-            )
-            current_step = await _submit_step(
-                submit_fn,
-                flow_id,
-                {"next_step_id": menu_choice},
-                client=client,
-                helper_type=helper_type,
-                last_menu_choice=last_menu_choice,
-                current_step=current_step,
-            )
-
-        elif result_type == _FlowType.FORM:
-            # _handle_form_step pops only the keys declared in the current
-            # step's data_schema, leaving any other keys in remaining_config
-            # for subsequent steps (HA can present multi-step forms, e.g.
-            # statistics: user step then pick-characteristic step), and records
-            # what it consumed for any later step that redeclares the same
-            # field.
-            saw_form_step = True
-            consumed_form_keys: set[str] = set()
-            form_data = _auto_confirm_form_payload(current_step)
-            if form_data is None:
-                form_data = _handle_form_step(
-                    flow_id,
-                    current_step,
-                    remaining_config,
-                    ignored_config_keys,
-                    consumed_form_keys,
-                    reuse_state,
-                )
-            if consumed_form_keys:
-                any_form_key_consumed = True
-            logger.debug(
-                f"Flow step {step_num}: form submit "
-                f"(step_id={current_step.get('step_id')}, keys={list(form_data.keys())})"
-            )
-            current_step = await _submit_step(
-                submit_fn,
-                flow_id,
-                form_data,
-                client=client,
-                helper_type=helper_type,
-                last_menu_choice=last_menu_choice,
-                current_step=current_step,
-            )
-
-        elif result_type in _UNDRIVABLE_STEP_TYPES:
-            raise_tool_error(
-                create_error_response(
-                    ErrorCode.SERVICE_CALL_FAILED,
-                    f"Flow reached a '{result_type}' step that cannot be "
-                    "completed via MCP (browser/OAuth authorization or an "
-                    "asynchronous provider step)",
-                    suggestions=[
-                        "Complete this flow in the Home Assistant UI "
-                        "(Settings > Devices & Services)."
-                    ],
-                    context={"flow_id": flow_id, "details": current_step},
-                )
-            )
-
-        else:
-            raise_tool_error(
-                create_error_response(
-                    ErrorCode.INTERNAL_UNEXPECTED,
-                    f"Unexpected flow result type: {result_type}",
-                    context={"flow_id": flow_id, "details": current_step},
-                )
-            )
-
-    raise_tool_error(
-        create_error_response(
-            ErrorCode.TIMEOUT_OPERATION,
-            f"Flow exceeded {max_steps} steps",
-            context={
-                "flow_id": flow_id,
-                "max_steps": max_steps,
-                "consumed_menu_selections": consumed_menu_selections,
-            },
-        )
-    )
-
-
-async def _handle_config_subentry_flow_steps(
-    client: Any,
-    flow_id: str,
-    initial_step: dict[str, Any],
-    config: dict[str, Any],
-    *,
-    is_reconfigure: bool,
-) -> dict[str, Any]:
-    """Walk a config subentry flow and accept HA's reconfigure-success abort.
-
-    Successful results include ``warnings`` when caller-supplied config keys
-    were not declared by any flow step, or when a key was resubmitted to a
-    later step that redeclared it. Fields redeclared by a later step are filled
-    on the same terms as :func:`_handle_flow_steps`.
-    """
-    remaining_config = dict(config)
-    current_step = initial_step
-    last_menu_choice: str | None = None
-    consumed_menu_selections: list[str] = []
-    consumed_menu_selection_keys: list[str] = []
-    ignored_config_keys: set[str] = set()
-    reuse_state = _ReuseState()
-    max_steps = _flow_step_budget(config)
-
-    for step_num in range(max_steps):
-        result_type = current_step.get("type")
-
-        if result_type == _FlowType.CREATE_ENTRY:
-            response: dict[str, Any] = {
-                "success": True,
-                "operation": "created",
-                "flow_result": current_step,
-            }
-            warnings = _success_warnings(
-                ignored_config_keys, remaining_config, reuse_state
-            )
-            if warnings:
-                response["warnings"] = warnings
-            return response
-
-        if result_type == _FlowType.ABORT:
-            reason = current_step.get("reason")
-            if is_reconfigure and reason in _RECONFIGURE_SUCCESS_REASONS:
-                response = {
-                    "success": True,
-                    "operation": "reconfigured",
-                    "flow_result": current_step,
-                }
-                warnings = _success_warnings(
-                    ignored_config_keys, remaining_config, reuse_state
-                )
-                if warnings:
-                    response["warnings"] = warnings
-                return response
-            raise_tool_error(
-                create_error_response(
-                    ErrorCode.SERVICE_CALL_FAILED,
-                    f"Config subentry flow aborted: {reason}",
-                    context={"flow_id": flow_id, "details": current_step},
-                )
-            )
-
-        if result_type == _FlowType.MENU:
-            menu_choice = _handle_menu_step(
-                flow_id,
-                current_step,
-                remaining_config,
-                consumed_menu_selections,
-                consumed_menu_selection_keys,
-            )
-            last_menu_choice = menu_choice
-            logger.debug(
-                "Config subentry flow step %s: menu %s (step_id=%s)",
-                step_num,
-                menu_choice,
-                current_step.get("step_id"),
-            )
-            current_step = await _submit_step(
-                client.submit_config_subentry_flow_step,
-                flow_id,
-                {"next_step_id": menu_choice},
-                client=client,
-                helper_type=None,
-                last_menu_choice=last_menu_choice,
-                current_step=current_step,
-            )
-            continue
-
-        if result_type == _FlowType.FORM:
-            form_data = _handle_form_step(
-                flow_id,
-                current_step,
-                remaining_config,
-                ignored_config_keys,
-                reuse_state=reuse_state,
-            )
-            logger.debug(
-                "Config subentry flow step %s: form submit (step_id=%s, keys=%s)",
-                step_num,
-                current_step.get("step_id"),
-                sorted(form_data.keys()),
-            )
-            current_step = await _submit_step(
-                client.submit_config_subentry_flow_step,
-                flow_id,
-                form_data,
-                client=client,
-                helper_type=None,
-                last_menu_choice=last_menu_choice,
-                current_step=current_step,
-            )
-            continue
-
-        if result_type in {"progress", "progress_done"}:
-            raise_tool_error(
-                create_error_response(
-                    ErrorCode.SERVICE_CALL_FAILED,
-                    "Config subentry flow requires an asynchronous progress step",
-                    suggestions=[
-                        "Complete the provider setup in Home Assistant so the external resource is available.",
-                        "Retry the same ha_config_set_helper call after the resource is ready.",
-                    ],
-                    context={"flow_id": flow_id, "details": current_step},
-                )
-            )
-
-        raise_tool_error(
-            create_error_response(
-                ErrorCode.INTERNAL_UNEXPECTED,
-                f"Unexpected config subentry flow result type: {result_type}",
-                context={"flow_id": flow_id, "details": current_step},
-            )
-        )
-
-    raise_tool_error(
-        create_error_response(
-            ErrorCode.TIMEOUT_OPERATION,
-            f"Config subentry flow exceeded {max_steps} steps",
-            context={
-                "flow_id": flow_id,
-                "max_steps": max_steps,
-                "consumed_menu_selections": consumed_menu_selections,
-            },
-        )
-    )
 
 
 async def set_config_subentry(
