+++ tests/src/unit/test_helper_schema_inline.py
@@ -34,10 +34,8 @@ from unittest.mock import AsyncMock
 import pytest
 from fastmcp.exceptions import ToolError
 
-from ha_mcp.tools.config_entry_flow import (
-    FLOW_HELPER_TYPES,
-    fetch_helper_flow_info,
-)
+from ha_mcp.tools.config_entry_flow import FLOW_HELPER_TYPES
+from ha_mcp.tools.config_entry_flow_walker import fetch_helper_flow_info
 from ha_mcp.tools.tools_config_helpers import (
     SIMPLE_HELPER_SCHEMAS,
     SIMPLE_HELPER_TYPES,
@@ -771,7 +769,7 @@ class TestExtractMenuChoiceFromConfig:
             ({"menu_option": "binary_sensor"}, "binary_sensor"),
             # Multiple keys: first in _MENU_CHOICE_CONFIG_KEYS order wins —
             # the same canonical order _handle_menu_step consumes by
-            # (_MENU_SELECTION_KEY_ORDER in config_entry_flow.py).
+            # (_MENU_SELECTION_KEY_ORDER in config_entry_flow_menu.py).
             (
                 {"group_type": "light", "next_step_id": "sensor"},
                 "light",
